requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1<0.16
gunicorn>=21.2,<22.0
python-stdnum>=1.19,<1.20
django-eav2>=1.5.0,<=1.6.0
django-cors-headers>=4.3.1,<=4.4.0
django-easy-audit>=1.3.6
reportlab>=4.2.0